[CHANGED] prefixed $ in amounts and added commas to amounts

+++ crowdfunding/forms.py
@@ -3,9 +3,7 @@ from crowdfunding.models import CrowdFund
 
 
 class PaymentForm(forms.Form):
-    amount = forms.DecimalField(
-        min_value=1, decimal_places=2, required=True
-    )
+    amount = forms.IntegerField(min_value=1, required=True)
     message = forms.CharField(max_length=250, required=False)
